Moved common use component to components folder
